Fixed SymbolTable entry for "round" Fixed arguments from INT to DOUBLE CLOSED - task 22: round Std Lib function takes an INT http://github.com/Hexcore/HxCAS/issues/issue/22

+++ src/com/hexcore/cas/rulesystems/SymbolTable.java
@@ -134,7 +134,7 @@ public class SymbolTable
 		roundEntry.type = TableEntry.Type.INT;
 		roundEntry.kind = TableEntry.Kind.METHOD;
 		roundEntry.name = "round";
-		roundEntry.arguments = new ArgList(TableEntry.Type.INT);
+		roundEntry.arguments = new ArgList(TableEntry.Type.DOUBLE);
 		insert(roundEntry);
 		
 		TableEntry countEntry = new TableEntry();
